feat(app): adjust union icons component
